feat: mover página para  e adicionar layout completo
